% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\alias{maketools_diagnostics}
\title{Diagnostics Report}
\usage{
maketools_diagnostics()
}
\description{
Print some diagnostics about your compiler environment. These are also
shown when the \code{maketools} package is attached.
}
\seealso{
Other maketools: 
\code{\link{make}()},
\code{\link{pkgconfig}},
\code{\link{r_config}},
\code{\link{sysdeps}}
}
\concept{maketools}
