% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{find_logo}
\alias{find_logo}
\title{Find logo}
\usage{
find_logo(path = ".")
}
\arguments{
\item{path}{root directory of package}
}
\description{
Match logic from pkgdown but return path relative package root.
}
