% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgconfig.R
\name{pc_info}
\alias{pc_info}
\alias{pkg_list}
\alias{pkg_exists}
\alias{pkg_version}
\alias{pkg_cflags}
\alias{pkg_libs}
\alias{pkg_info}
\title{Query pkg-config}
\usage{
pc_info()

pkg_list()

pkg_exists(pkg = "libcurl")

pkg_version(pkg = "libcurl")

pkg_cflags(pkg = "libcurl")

pkg_libs(pkg = "libcurl", static = FALSE)

pkg_info(pkg = "libcurl")
}
\arguments{
\item{pkg}{names of the pkg-config libraries to query}

\item{static}{get libs for static linking, i.e. include dependencies}
}
\description{
Wrappers for the pkg-config utility to query information on C/C++
libraries installed on your system.
}
\examples{
# Check if pkg-config is available
pc_info()
pkg_list()
}
