% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-compile.R
\name{r-compile}
\alias{r-compile}
\alias{r_compile}
\title{Compile source files}
\usage{
r_compile(src, obj = NULL)
}
\arguments{
\item{src}{path to source file}

\item{obj}{optional name of output file}
}
\description{
Compile C / C++ / Fortran source files using the compiler configured
by R Makeconf.
}
\details{
This program literally calls `make yourfile.o -f /path/to/R/Makeconf`.
This is exacltly what R does when building packages and hence the best
way to test if your compiler is working.
}
\examples{
# Test the CXX compiler
testprog <- '#include <iostream>\\nint main() {std::cout << "Hello World!";}'
writeLines(testprog, con = 'test.cc')
r_compile('test.cc')
}
