% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-config.R
\name{r_cmd_config}
\alias{r_cmd_config}
\alias{r_has_cc}
\alias{r_has_cxx}
\alias{r_has_cxx11}
\alias{r_has_fc}
\alias{r_has}
\title{R CMD Config}
\usage{
r_has_cc()

r_has_cxx()

r_has_cxx11()

r_has_fc()

r_cmd_config(VAR = "--all")

r_has(VAR)
}
\arguments{
\item{VAR}{value passed to `R CMD config` such as `CXX` or `FC`}
}
\description{
Cross-platform wrappers for `R CMD config` to lookup the availability
of the compiler.
}
