% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-config.R
\name{r_config}
\alias{r_config}
\alias{cc_info}
\alias{cxx_info}
\alias{cxx11_info}
\alias{cxx14_info}
\alias{cxx17_info}
\alias{fc_info}
\title{R CMD Config}
\usage{
cc_info()

cxx_info()

cxx11_info()

cxx14_info()

cxx17_info()

fc_info()
}
\description{
Cross-platform wrappers for \code{R CMD config} to lookup the availability
of the compiler.
}
\examples{
# This runs 'R CMD CONFIG CXX'
r_cmd_config("CXX")

# Show C++ config:
cxx_info()
}
