% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{rtools_find}
\alias{rtools_find}
\alias{rtools_install}
\title{Find Rtools}
\usage{
rtools_find()

rtools_install(silent = FALSE)
}
\arguments{
\item{silent}{perform automatic unattended installation (answer YES
to all questions)}
}
\description{
Lookup and test installations of Rtools in Windows registry.
}
