% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdeps.R
\name{sysdeps}
\alias{sysdeps}
\alias{dpkg_sysdeps}
\title{Find System Dependencies}
\usage{
dpkg_sysdeps(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{name of an installed packages}

\item{lib.loc}{path to the library of the package}
}
\description{
Check which external libraries that an installed package links to.
Currently only works on debian/ubuntu.
}
