% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgconfig.R
\name{pkgconfig}
\alias{pkgconfig}
\alias{pc_info}
\alias{pc_pkg_list}
\alias{pc_pkg_exists}
\alias{pc_pkg_version}
\alias{pc_pkg_cflags}
\alias{pc_pkg_libs}
\alias{pc_pkg_info}
\title{Query pkg-config}
\usage{
pc_info()

pc_pkg_list()

pc_pkg_exists(pkg = "libcurl")

pc_pkg_version(pkg = "libcurl")

pc_pkg_cflags(pkg = "libcurl")

pc_pkg_libs(pkg = "libcurl", static = FALSE)

pc_pkg_info(pkg = "libcurl")
}
\arguments{
\item{pkg}{names of the pkg-config libraries to query}

\item{static}{get libs for static linking, i.e. include dependencies}
}
\description{
Wrappers for the pkg-config utility to query information on C/C++ libraries
that are available on your system.
}
\examples{
# Check if pkg-config is available
(info <- pc_info())
if(info$available)
  pc_pkg_list()
}
\seealso{
Other maketools: 
\code{\link{diagnostics}},
\code{\link{make}()},
\code{\link{r_config}},
\code{\link{sysdeps}}
}
\concept{maketools}
