% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-compile.R
\name{r-compile}
\alias{r-compile}
\alias{make}
\title{Compile source files}
\usage{
make(target = "all", makefile = r_makeconf_path())
}
\arguments{
\item{target}{name of output file that you want to make}

\item{makefile}{path to the \code{Makefile}. Defaults to the \code{Makeconf} which
R uses when building R packages.}
}
\description{
Compile C / C++ / Fortran source files using the compiler configured
by R Makeconf.
}
\details{
This program literally calls \code{make yourfile.o -f /path/to/R/Makeconf}.
This is exacltly what R does when building packages and hence the best
way to test if your compiler is working.
}
\examples{
# Test the CXX compiler
testprog <- '#include <iostream>\\nint main() {std::cout << "Hello World!";}'
writeLines(testprog, con = 'testprog.cc')
make('testprog')
system('./testprog')
}
