% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-makeconf.R
\name{r_cmd_config}
\alias{r_cmd_config}
\alias{make_call}
\alias{make_info}
\title{Check your makeconf}
\usage{
r_cmd_config(VAR = "--all")

make_call(cmd = "$(CC)", args = "--version")

make_info()
}
\arguments{
\item{VAR}{value passed to `R CMD config` such as `CXX` or `FC`}

\item{cmd}{command to invoke (may be a variable)}

\item{args}{additional arguments for `cmd`}
}
\description{
Wrappers for `make` and `R CMD config` to lookup and test make variables
from your local `Makeconf` file.
}
\examples{
make_call('$(CXX)', '--version')
# Where your makeconf is stored:
make_info()
}
