% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-makeconf.R
\name{r_cmd_config}
\alias{r_cmd_config}
\alias{r_make_test}
\title{Check your makeconf}
\usage{
r_cmd_config(VAR = "--all")

r_make_test(VAR, args = "--version")
}
\arguments{
\item{VAR}{value passed to `R CMD config` such as `CXX` or `FC`}

\item{args}{argument(s) to test program}
}
\description{
Wrappers for `make` and `R CMD config` to lookup and test make variables
from your local `Makeconf` file.
}
