% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeconf.R
\name{r_cmd_config}
\alias{r_cmd_config}
\alias{make_call}
\alias{make_echo}
\alias{make_info}
\title{Check your makeconf}
\usage{
r_cmd_config(VAR = "--all")

make_call(cmd = "$(CC)", args = "--version")

make_echo(VAR = "$(CC)")

make_info()
}
\arguments{
\item{VAR}{value passed to \code{R CMD config} such as \code{CXX} or \code{FC}}

\item{cmd}{command to invoke (may be a variable)}

\item{args}{additional arguments for \code{cmd}}
}
\description{
Wrappers for \code{make} and \code{R CMD config} to lookup and test make variables
from your local \code{Makeconf} file.
}
\examples{
make_call('$(CXX)', '--version')
# Where your makeconf is stored:
make_info()
}
