% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{r_config}
\alias{r_config}
\alias{cc_info}
\alias{cxx_info}
\alias{cxx11_info}
\alias{cxx14_info}
\alias{cxx17_info}
\alias{fc_info}
\alias{r_cmd_config}
\title{R CMD Config}
\usage{
cc_info()

cxx_info()

cxx11_info()

cxx14_info()

cxx17_info()

fc_info()

r_cmd_config(VAR = "--all")
}
\arguments{
\item{VAR}{value passed to \verb{R CMD config} such as \code{CXX} or \code{FC}}
}
\description{
Cross-platform wrappers for \verb{R CMD config} to lookup the availability
of the compiler.
}
\examples{
# This runs 'R CMD CONFIG CXX'
r_cmd_config("CXX")

# Show C++ config:
cxx_info()
}
\seealso{
Other maketools: 
\code{\link{diagnostics}},
\code{\link{make}()},
\code{\link{pkgconfig}},
\code{\link{sysdeps}}
}
\concept{maketools}
