% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-config.R
\name{r_config}
\alias{r_config}
\alias{r_test_cc}
\alias{r_test_cxx}
\alias{r_test_cxx11}
\alias{r_test_fc}
\title{R CMD Config}
\usage{
r_test_cc()

r_test_cxx()

r_test_cxx11()

r_test_fc()
}
\description{
Cross-platform wrappers for `R CMD config` to lookup the availability
of the compiler.
}
\examples{
# This runs 'R CMD CONFIG CXX'
r_cmd_config("CXX")

# Show default C++ compiler:
r_test_cxx()
}
