% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{rtools}
\alias{rtools}
\alias{rtools_info}
\alias{rtools_setup}
\alias{rtools_install}
\title{Setup Rtools}
\usage{
rtools_info()

rtools_setup()

rtools_install(silent = FALSE)
}
\arguments{
\item{silent}{perform automatic unattended installation (answer YES
to all questions)}
}
\description{
Windows does not ship with a native compiler toolchain. In order to build
R packages with compiled code, we need the R for Windows toolchain bundle
called \href{https://cran.r-project.org/bin/windows/Rtools/}{Rtools}. Below are
utilities to check if and where Rtools is installed, and to set the proper
\code{PATH} and \code{BINPREF} variables to make it work.
}
