% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{rtools_info}
\alias{rtools_info}
\alias{rtools_install}
\alias{rtools_setup}
\title{Find Rtools}
\usage{
rtools_info()

rtools_install(silent = FALSE)

rtools_setup()
}
\arguments{
\item{silent}{perform automatic unattended installation (answer YES
to all questions)}
}
\description{
Lookup and test installations of Rtools in Windows registry.
}
