% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdeps.R
\name{sysdeps}
\alias{sysdeps}
\alias{package_sysdeps}
\alias{package_sysdeps_string}
\title{Package System Dependencies}
\usage{
package_sysdeps(pkg, lib.loc = NULL)

package_sysdeps_string(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{name of an installed packages}

\item{lib.loc}{path to the library of the package}
}
\description{
Finds the shared libraries that an installed R package links to by running
\code{ldd} on the package \code{so} file. Then uses system package manager (i.e. \code{dpkg}
or \code{rpm}) to locate the rpm/deb package that contains the shared libraries,
headers, and sources for this library.
}
\details{
For common Linux distributions, the output also includes a URL to the homepage
of this package within the given Linux distribution. Here we can typically find
more information about the library, such as configuration options, dependencies,
and patches applied by the distribution.

This tool shows what a package links to, i.e. the run-time dependencies of an
installed R package. This is particularly important to know if you want to
distribute a pre-compiled binary of the R package. Note that this tool does not
include dependencies that are only needed at build-time, such as static or
header-only libraries.
}
