% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdeps.R
\name{sysdeps}
\alias{sysdeps}
\alias{dpkg_sysdeps}
\alias{dpkg_sysdeps_string}
\title{Show System Dependencies}
\usage{
dpkg_sysdeps(pkg, lib.loc = NULL)

dpkg_sysdeps_string(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{name of an installed packages}

\item{lib.loc}{path to the library of the package}
}
\description{
Finds the shared libraries that an installed package links to by running \code{ldd}
on the package \code{so} file. Then uses \code{dpkg} find the debian packages that contain
the libs and the headers for this library.
}
