% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdeps.R
\name{sysdeps}
\alias{sysdeps}
\alias{package_sysdeps}
\alias{package_sysdeps_string}
\alias{package_links_to}
\title{Package System Dependencies}
\usage{
package_sysdeps(pkg, lib.loc = NULL)

package_sysdeps_string(pkg, lib.loc = NULL)

package_links_to(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{name of an installed R package}

\item{lib.loc}{path to the R package directory for this package}
}
\description{
Finds the shared libraries that an installed R package links to by running
\code{ldd} on the package \code{so} file. Then uses system package manager (e.g. \code{dpkg}
or \code{rpm}) to locate the rpm/deb package that contains the shared libraries,
headers, and sources for this library.
}
\details{
For common Linux distributions, the output also includes a URL to the homepage
of this package within the given Linux distribution. Here we can typically find
more information about the package, such as configuration options, dependencies,
and custom patches applied by the distribution maintainers.

Because we use \code{ldd}, this only shows hard run-time dependencies of an installed
R package. It does not show dependencies that are only needed at build-time, such
as static or header-only libraries.
}
